updated framework, fixed some bugs

+++ BMBH_Views/RoleMgr.aspx.cs
@@ -1,27 +1,24 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using ColorCode;
+using HtmlAgilityPack;
+using System;
+using System.Configuration;
+using System.Data;
+using System.Data.SqlClient;
+using System.Text;
 using System.Web;
 using System.Web.UI;
 using System.Web.UI.WebControls;
-using System.Data;
-using System.Configuration;
-using System.Data.SqlClient;
-using ColorCode;
-using System.Text.RegularExpressions;
-using HtmlAgilityPack;
-using System.Text;
 
-namespace BMBH_View
+namespace BMBHviews
 {
     public partial class UserMan : System.Web.UI.Page
     {
-        static string prevPage = String.Empty;
+        //private static readonly string prevPage = string.Empty;
 
         public static string Base64Decode(string base64EncodedData)
         {
-            var base64EncodedBytes = System.Convert.FromBase64String(base64EncodedData);
-            return System.Text.Encoding.UTF8.GetString(base64EncodedBytes);
+            byte[] base64EncodedBytes = Convert.FromBase64String(base64EncodedData);
+            return Encoding.UTF8.GetString(base64EncodedBytes);
         }
 
         protected void Page_Load(object sender, EventArgs e)
@@ -29,86 +26,92 @@ namespace BMBH_View
             if (!Page.IsPostBack) // on first load
             {
                 cboRole.SelectedValue = (string)Session["RoleId"];
-//                SqlDataSource2.SelectCommand = "GetPermittedViewsByRole @RoleId,''";
+                //                SqlDataSource2.SelectCommand = "GetPermittedViewsByRole @RoleId,''";
             }
 
-            if (this.Request["__EVENTARGUMENT"] == "PostFromNew_Send") // add new role
+            if (Request["__EVENTARGUMENT"] == "PostFromNew_Send") // add new role
             {
                 SQLexecute("insert into Roles (RoleName) VALUES ('" + txtRoleName.Text.Trim() + "')");
                 Response.Redirect("RoleMgr.aspx");
             }
 
-            if (this.Request["__EVENTARGUMENT"] == "EditView_OK") // confirm edit view
+            if (Request["__EVENTARGUMENT"] == "EditView_OK") // confirm edit view
             {
                 string sViewDefinition = Server.UrlDecode(Base64Decode(HiddenField1.Value));
                 SQLexecute(GetTextOnly(sViewDefinition));
                 //Response.Redirect("RoleMgr.aspx");
             }
 
-            if (this.Request["__EVENTTARGET"] == updViews.ClientID) // search views by filter textbox
+            if (Request["__EVENTTARGET"] == updViews.ClientID) // search views by filter textbox
             {
-                string value = this.Request["__EVENTARGUMENT"];
+                string value = Request["__EVENTARGUMENT"];
                 SqlDataSource2.SelectCommand = "GetPermittedViewsByRole @RoleId,'" + value + "'";
                 dgdViewPermissions.DataBind();
             }
             else
-                SqlDataSource2.SelectCommand = "GetPermittedViewsByRole @RoleId,''";
-        }
-
-        private void SQLexecute_async(string sSQL)
-        {
-            // Given command text and connection string, asynchronously execute
-            // the specified command against the connection. For this example,
-            // the code displays an indicator as it is working, verifying the 
-            // asynchronous behavior. 
-            string sConnString = ConfigurationManager.ConnectionStrings["BMBHViewsConnectionString"].ConnectionString;
-            using (SqlConnection connection = new SqlConnection(sConnString))
             {
-                try
-                {
-                    //int count = 0;
-                    SqlCommand command = new SqlCommand(sSQL, connection);
-                    command.CommandTimeout = 300;
-                    connection.Open();
-
-                    IAsyncResult result = command.BeginExecuteNonQuery();
-                    while (!result.IsCompleted)
-                    {
-                        //Console.WriteLine("Waiting ({0})", count++);
-                        // Wait for 1/10 second, so the counter
-                        // does not consume all available resources 
-                        // on the main thread.
-                        System.Threading.Thread.Sleep(100);
-                    }
-                    //Console.WriteLine("Command complete. Affected {0} rows.",
-                    command.EndExecuteNonQuery(result);
-                }
-                catch (SqlException ex)
-                {
-                    ShowMsg(ex.Message);
-                }
-                catch (InvalidOperationException ex)
-                {
-                    ShowMsg(ex.Message);
-                }
-                catch (Exception ex)
-                {
-                    // You might want to pass these errors
-                    // back out to the caller.
-                    ShowMsg(ex.Message);
-                }
+                SqlDataSource2.SelectCommand = "GetPermittedViewsByRole @RoleId,''";
             }
         }
 
-        private void SQLexecute(string sSQL)
+        //private void SQLexecute_async(string sSQL)
+        //{
+        //    // Given command text and connection string, asynchronously execute
+        //    // the specified command against the connection. For this example,
+        //    // the code displays an indicator as it is working, verifying the 
+        //    // asynchronous behavior. 
+        //    string sConnString = ConfigurationManager.ConnectionStrings["BMBHViewsConnectionString"].ConnectionString;
+        //    using (SqlConnection connection = new SqlConnection(sConnString))
+        //    {
+        //        try
+        //        {
+        //            //int count = 0;
+        //            SqlCommand command = new SqlCommand(sSQL, connection)
+        //            {
+        //                CommandTimeout = 300
+        //            };
+        //            connection.Open();
+
+        //            IAsyncResult result = command.BeginExecuteNonQuery();
+        //            while (!result.IsCompleted)
+        //            {
+        //                //Console.WriteLine("Waiting ({0})", count++);
+        //                // Wait for 1/10 second, so the counter
+        //                // does not consume all available resources 
+        //                // on the main thread.
+        //                System.Threading.Thread.Sleep(100);
+        //            }
+        //            //Console.WriteLine("Command complete. Affected {0} rows.",
+        //            command.EndExecuteNonQuery(result);
+        //        }
+        //        catch (SqlException ex)
+        //        {
+        //            ShowMsg(ex.Message);
+        //        }
+        //        catch (InvalidOperationException ex)
+        //        {
+        //            ShowMsg(ex.Message);
+        //        }
+        //        catch (Exception ex)
+        //        {
+        //            // You might want to pass these errors
+        //            // back out to the caller.
+        //            ShowMsg(ex.Message);
+        //        }
+        //    }
+        //}
+
+        private static void SQLexecute(string sSQL)
         {
-            String sConnString = ConfigurationManager.ConnectionStrings["BMBHViewsConnectionString"].ConnectionString;
+            string sConnString = ConfigurationManager.ConnectionStrings["BMBHViewsConnectionString"].ConnectionString;
             SqlConnection con = new SqlConnection(sConnString);
-            SqlCommand cmd = new SqlCommand();
-            cmd.CommandType = CommandType.Text;
-            cmd.CommandText = sSQL;
-            cmd.Connection = con;
-            cmd.CommandTimeout = 900;
+            SqlCommand cmd = new SqlCommand
+            {
+                CommandType = CommandType.Text,
+                CommandText = sSQL,
+                Connection = con,
+                CommandTimeout = 900
+            };
 
             try
             {
@@ -118,7 +121,7 @@ namespace BMBH_View
             }
             catch (Exception ex)
             {
-                throw ex;
+                throw;
             }
             finally
             {
@@ -130,32 +133,40 @@ namespace BMBH_View
         private void ShowMsg(string message)
         {
             message = "alert('" + message + "')";
-            ScriptManager.RegisterClientScriptBlock((Page as Control), this.GetType(), "alert", message, true);
+            ScriptManager.RegisterClientScriptBlock((Page as Control), GetType(), "alert", message, true);
         }
 
         protected void chkPermission_CheckedChanged(object sender, EventArgs e)
         {
+            if (e == null)
+                throw new ArgumentNullException(nameof(e));
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             CheckBox chk = (CheckBox)sender;
             GridViewRow row = (GridViewRow)chk.NamingContainer;
-            var sViewName = row.Cells[0].Text;
+            string sViewName = row.Cells[0].Text;
             string sSQL;
 
             if (chk.Checked)
             {
                 sSQL = "EXEC SetRolePermission " + cboRole.SelectedValue + ",'" + sViewName + "'";
-                try { SQLexecute(sSQL); } catch (Exception ex) { ShowMsg(ex.ToString());  }
+                try { SQLexecute(sSQL); } catch (Exception ex) { ShowMsg(ex.ToString()); }
                 sSQL = "EXEC SetRolePermission " + cboRole.SelectedValue + ",'" + sViewName + "'"; // hack to prevent from clicking two times
                 try { SQLexecute(sSQL); } catch (Exception ex) { ShowMsg(ex.ToString()); }
             }
             else
             {
                 sSQL = "EXEC UnsetRolePermission " + cboRole.SelectedValue + ",'" + sViewName + "'";
-                try { SQLexecute(sSQL); } catch (Exception ex) { }
+                try { SQLexecute(sSQL); } catch (Exception) { }
             }
         }
 
         protected void btnShowView_Click(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             Button btn = (Button)sender;
             GridViewRow row = (GridViewRow)btn.NamingContainer;
             Session["View"] = row.Cells[0].Text;
@@ -166,6 +177,9 @@ namespace BMBH_View
 
         protected void btnValueCnt_Click(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             Button btn = (Button)sender;
             GridViewRow row = (GridViewRow)btn.NamingContainer;
             string sView = row.Cells[0].Text;
@@ -176,10 +190,13 @@ namespace BMBH_View
         // OK button to confirm changes on button text and selected panel
         protected void btnOK_Click(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             GridViewRow row = (GridViewRow)((Button)sender).NamingContainer;
             string sCaption = ((TextBox)row.FindControl("txtCaption")).Text;
             string sView = row.Cells[0].Text;
-            
+
             SQLexecute("EXEC AddViewToPanel '" + sView + "','" + sCaption + "'");
             SQLexecute("EXEC AddViewToPanel '" + sView + "','" + sCaption + "'"); // hack to prevent from clicking two times
             //Server.Transfer("RoleMgr.aspx");
@@ -187,6 +204,9 @@ namespace BMBH_View
 
         protected void btnOK_Sorter_Click(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             GridViewRow row = (GridViewRow)((Button)sender).NamingContainer;
             string sSorter = ((TextBox)row.FindControl("txtSorter")).Text;
             string sView = row.Cells[0].Text;
@@ -197,6 +217,9 @@ namespace BMBH_View
 
         protected void cboPanel_SelectedIndexChanged(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             GridViewRow row = (GridViewRow)((DropDownList)sender).NamingContainer;
             string sPanel = ((DropDownList)row.FindControl("cboPanel")).SelectedValue;
             string sView = row.Cells[0].Text;
@@ -215,14 +238,14 @@ namespace BMBH_View
         {
         }
 
-        private string SQLexecute_SingleResult(string sSQL)
+        private static string SQLexecute_SingleResult(string sSQL)
         {
-            String sConnString = ConfigurationManager.ConnectionStrings["BMBHViewsConnectionString"].ConnectionString;
+            string sConnString = ConfigurationManager.ConnectionStrings["BMBHViewsConnectionString"].ConnectionString;
             SqlConnection con = new SqlConnection(sConnString);
-            using (var command = new SqlCommand(sSQL, con))
+            using (SqlCommand command = new SqlCommand(sSQL, con))
             {
                 con.Open();
-                using (var reader = command.ExecuteReader())
+                using (SqlDataReader reader = command.ExecuteReader())
                 {
                     reader.Read();
                     return reader.GetString(0);
@@ -232,6 +255,9 @@ namespace BMBH_View
 
         protected void btnEditView_Click(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             GridViewRow row = (GridViewRow)((Button)sender).NamingContainer;
             string sView = row.Cells[0].Text;
             string sSQL = "SELECT OBJECT_DEFINITION (OBJECT_ID(N'" + sView + "'))";
@@ -239,7 +265,7 @@ namespace BMBH_View
             sResult = new CodeColorizer().Colorize(sResult, Languages.Sql);
             sResult = sResult.Replace("<div style=\"color:Black;background-color:White;\">", "").Replace("</div>", "");
             txtViewDefinition.InnerHtml = sResult;
-      
+
             MPE_EditView.Show();
         }
 
@@ -266,15 +292,21 @@ namespace BMBH_View
         // update lookup values for selected view
         protected void btnUpdateLookups_Click(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             GridViewRow row = (GridViewRow)((Button)sender).NamingContainer;
             string sView = row.Cells[0].Text;
-            ScriptManager.RegisterClientScriptBlock((Page as Control), this.GetType(), "ShowHourglass", "document.body.style.cursor = 'progress';", true);
+            ScriptManager.RegisterClientScriptBlock((Page as Control), GetType(), "ShowHourglass", "document.body.style.cursor = 'progress';", true);
             SQLexecute("EXEC CreateLookups '" + sView + "'");
-            ScriptManager.RegisterClientScriptBlock((Page as Control), this.GetType(), "HideHourglass", "document.body.style.cursor = 'default';", true);
+            ScriptManager.RegisterClientScriptBlock((Page as Control), GetType(), "HideHourglass", "document.body.style.cursor = 'default';", true);
         }
 
         protected void chkUseLookups_CheckedChanged(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             GridViewRow row = (GridViewRow)((CheckBox)sender).NamingContainer;
             string sView = row.Cells[0].Text;
             string sUseLookup = ((CheckBox)sender).Checked ? "1" : "0";
@@ -283,6 +315,9 @@ namespace BMBH_View
 
         protected void btnDelView_Click(object sender, EventArgs e)
         {
+            if (sender == null)
+                throw new ArgumentNullException(nameof(sender));
+
             GridViewRow row = (GridViewRow)((Button)sender).NamingContainer;
             string sView = row.Cells[0].Text;
             SQLexecute("delete from RoleViews where ViewName='" + sView + "'");
